test: fold image_ref e2e cases into the existing realtime suite
