+++ packages/sdk/tests/e2e-realtime.test.ts
@@ -118,4 +118,85 @@ describe.concurrent("Realtime E2E Tests", { timeout: TIMEOUT, retry: 2 }, () => 
       expect(realtimeClient?.getConnectionState()).toBe("disconnected");
     });
   }
+
+  // POST /v1/files + image_ref via realtime. Uses lucy-2.1 to actually exercise
+  // upstream forwarding through the bouncer's image_ref prologue.
+  describe("Files API + image_ref", () => {
+    async function pngBlob(width: number, height: number, color = "#3070C0"): Promise<Blob> {
+      const canvas = document.createElement("canvas");
+      canvas.width = width;
+      canvas.height = height;
+      const ctx = canvas.getContext("2d");
+      if (!ctx) throw new Error("2D canvas context unavailable");
+      ctx.fillStyle = color;
+      ctx.fillRect(0, 0, width, height);
+      return new Promise((resolve, reject) => {
+        canvas.toBlob((blob) => (blob ? resolve(blob) : reject(new Error("toBlob failed"))), "image/png");
+      });
+    }
+
+    it("upload → get → delete round-trip", async () => {
+      const ref = await client.files.upload(await pngBlob(96, 96));
+      expect(ref.id).toMatch(/^file_/);
+      expect(ref.mime_type).toBe("image/png");
+
+      const fetched = await client.files.get(ref.id);
+      expect(fetched.id).toBe(ref.id);
+
+      await client.files.delete(ref.id);
+      await expect(client.files.get(ref.id)).rejects.toThrow(/Failed to get file/);
+    });
+
+    it("connects with initialState.image = ref.id", async () => {
+      const model = models.realtime("lucy-2.1");
+      const ref = await client.files.upload(await pngBlob(model.width, model.height));
+      const syntheticStream = createSyntheticStream(model.width, model.height);
+
+      let remoteStreamReceived = false;
+      let realtimeClient: Awaited<ReturnType<typeof client.realtime.connect>> | undefined;
+      const errors: DecartSDKError[] = [];
+
+      try {
+        realtimeClient = await client.realtime.connect(syntheticStream.stream, {
+          model,
+          initialState: { image: ref.id, prompt: { text: "make it cinematic", enhance: false } },
+          onRemoteStream: () => {
+            remoteStreamReceived = true;
+          },
+        });
+        realtimeClient.on("error", (err) => errors.push(err));
+
+        expect(["connected", "generating"]).toContain(realtimeClient.getConnectionState());
+        await waitUntil(() => remoteStreamReceived, "Timed out waiting for remote stream via image_ref");
+        expect(errors).toEqual([]);
+      } finally {
+        realtimeClient?.disconnect();
+        syntheticStream.stop();
+        await client.files.delete(ref.id).catch(() => {});
+      }
+    });
+
+    it("rt.set({ image: ref.id }) swaps the reference image mid-session", async () => {
+      const model = models.realtime("lucy-2.1");
+      const [refA, refB] = await Promise.all([
+        client.files.upload(await pngBlob(model.width, model.height, "#3070C0")),
+        client.files.upload(await pngBlob(model.width, model.height, "#C03070")),
+      ]);
+      const syntheticStream = createSyntheticStream(model.width, model.height);
+      let realtimeClient: Awaited<ReturnType<typeof client.realtime.connect>> | undefined;
+
+      try {
+        realtimeClient = await client.realtime.connect(syntheticStream.stream, {
+          model,
+          initialState: { image: refA.id, prompt: { text: "anime style", enhance: false } },
+          onRemoteStream: () => {},
+        });
+        await realtimeClient.set({ image: refB.id, prompt: "noir" });
+      } finally {
+        realtimeClient?.disconnect();
+        syntheticStream.stop();
+        await Promise.allSettled([client.files.delete(refA.id), client.files.delete(refB.id)]);
+      }
+    });
+  });
 });
