--- packages/sdk/vitest.config.e2e-image-ref.ts
@@ -1,17 +1,0 @@
-import { playwright } from "@vitest/browser-playwright";
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  define: {
-    __DECART_API_KEY__: JSON.stringify(process.env.DECART_API_KEY),
-  },
-  test: {
-    include: ["tests/e2e-realtime-image-ref.test.ts"],
-    browser: {
-      enabled: true,
-      provider: playwright(),
-      headless: true,
-      instances: [{ browser: "chromium" }],
-    },
-  },
-});
